--- examples/client-side-validation/app/routes/index.tsx
@@ -1,9 +1,0 @@
-import { redirect } from "remix";
-
-export function loader() {
-  return redirect('/example');
-}
-
-export default function Index() {
-  return;
-}
